Bellman Ford alghorithm

+++ models/graph/edge.go
@@ -1,20 +1,20 @@
 package graph
 
 type Edge struct {
-	weight      uint64
+	weight      float64
 	startVertex *Vertex
 	target      *Vertex
 }
 
-func NewEdge(targetVertex *Vertex, weight uint64) *Edge {
+func NewEdge(targetVertex *Vertex, weight float64) *Edge {
 	return &Edge{weight: weight, target: targetVertex}
 }
 
-func NewEdgeWithStartVertex(startVertex, targetVertex *Vertex, weight uint64) *Edge {
+func NewEdgeWithStartVertex(startVertex, targetVertex *Vertex, weight float64) *Edge {
 	return &Edge{weight: weight, startVertex: startVertex, target: targetVertex}
 }
 
-func (e *Edge) GetWeight() uint64 {
+func (e *Edge) GetWeight() float64 {
 	return e.weight
 }
